Made the Player Grapple to Grapple Point (made Physics work)
The issue was there was an if statement in the grappling rope script where an if statement was always false. This was caused by it testing if the grapple point and the final part of the line renderer was the same position. This never was the case and was always off by between 1 and 3. I fixed it by testing if the line no longer changes instead.