Added Functionality of Stalactite Not Killing Player The First Time it Falls On Player
When stalactite falls on the player, will put the player in a damaged state. While in this state the player moves at half the speed, can't jump and can't grapple. If another stalactite falls on the player during this time, the player dies. Visual effect will be added a bit later